Fixed some stuff, check description for more info
There's no reason to assume that the user is only on Debian, Arch, Fedora, or Ubuntu.  What if the user is on OpenSUSE? 
Debian has all those packages, so I added that in.
The user is supposed to install before continuing if they are on an unsupported distro, but 3 seconds is not enough time to do that.  Replaced the call to `sleep` with a prompt.